+++ CHANGELOG.md
@@ -2,4 +2,16 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.2.0](https://github.com/Puzzlout/static.website.boilerplate/compare/v1.0.1...v1.2.0) (2020-04-13)
+
+
+### Features
+
+* add svg loaders ([#9](https://github.com/Puzzlout/static.website.boilerplate/issues/9)) ([d162c19](https://github.com/Puzzlout/static.website.boilerplate/commit/d162c193261c8e480913e0a04264d821de6a10b4))
+* clean up language parse ([#3](https://github.com/Puzzlout/static.website.boilerplate/issues/3)) ([e18427b](https://github.com/Puzzlout/static.website.boilerplate/commit/e18427bdba480af86324f002176f496de1247517))
+* finalize language selection ([#3](https://github.com/Puzzlout/static.website.boilerplate/issues/3)) ([cc6f5e9](https://github.com/Puzzlout/static.website.boilerplate/commit/cc6f5e902be69d8711246b47360eb3eb8ac0b848))
+* force to define a default language in the Google Sheet ([#3](https://github.com/Puzzlout/static.website.boilerplate/issues/3)) ([5bc2bc6](https://github.com/Puzzlout/static.website.boilerplate/commit/5bc2bc6697acc8ab98db0821d58b194f745421bc))
+* update fallback text to svg ([#9](https://github.com/Puzzlout/static.website.boilerplate/issues/9)) ([7af05ab](https://github.com/Puzzlout/static.website.boilerplate/commit/7af05abdc7f5f23437a768c7a00e5fed5f1cd105))
+* update layout ([#8](https://github.com/Puzzlout/static.website.boilerplate/issues/8)) ([e01cf6f](https://github.com/Puzzlout/static.website.boilerplate/commit/e01cf6f9b020ea72869afc4eb043a789e7b7572b))
+
 ### [1.1.1](https://github.com/Puzzlout/static.website.boilerplate/compare/v1.0.1...v1.1.1) (2020-04-12)